Support hierarchical keys.
The docs here https://docs.microsoft.com/en-us/aspnet/core/security/key-vault-configuration state that using double hyphen "--" in keyvault secret keys will allow for hierarchical keys to be referenced. This doesn't seem to work. I'm not sure if there's something that's intended to be in AzureKeyVaultConfigurationProvider or in a IKeyVaultSecretManager implementation to handle this but I couldn't find anything.